requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
ConnPlotter==0.7a0
csa==0.1.8
cycler==0.10.0
decorator==4.4.1
enum-compat==0.0.3
enum34==1.1.6
future==0.18.2
gym==0.15.4
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazyarray==0.3.3
lxml==4.4.2
matplotlib==3.0.3
more-itertools==8.0.2
neo==0.6.1
numpy==1.17.4
opencv-python==4.1.2.30
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pyglet==1.3.2
Pygments==2.5.2
PyNEST==2.16.0
PyNN==0.9.5
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
quantities==0.12.3
requests==2.22.0
scipy==1.3.3
six==1.13.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
spalloc==2.0.1
SpiNNaker-DataSpecification==1!5.1.0
SpiNNaker-PACMAN==1!5.1.0
SpiNNFrontEndCommon==1!5.1.0
SpiNNMachine==1!5.1.0
SpiNNMan==1!5.1.0
SpiNNStorageHandlers==1!5.1.0
SpiNNUtilities==1!5.1.0
sPyNNaker==1!5.1.0
sPyNNaker8==1!5.1.0
Topology==2.16.0
torch==1.3.1
tornado==6.0.3
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
